requirements.txt
SQLAlchemy~=2.0.42
python-dotenv~=1.1.1
aiogram~=3.21.0
aiohttp~=3.12.15